normalized input and removed y coordinate

+++ inet4.4/src/inet/mobility/RL/InputState.h
@@ -13,11 +13,11 @@ typedef struct InputState {
     double latestPacketRSSI;      // RSSI value for the latest packet
     double latestPacketSNIR;      // SNIR value for the latest packet
     omnetpp::simtime_t latestPacketTimestamp;  // Timestamp of the latest packet
-    int numReceivedPackets;
+    double numReceivedPackets;
     omnetpp::simtime_t currentTimestamp;
     Coord coord;
 
-    InputState() : latestPacketRSSI(0.0), latestPacketSNIR(0.0), latestPacketTimestamp(0.0), numReceivedPackets(0), coord(0, 0) {} // Default constructor
+    InputState() : latestPacketRSSI(0.0), latestPacketSNIR(0.0), latestPacketTimestamp(0.0), numReceivedPackets(0.0), coord(0, 0) {} // Default constructor
 } InputState;
 
 }
